Video 2 Edited code
In this video I learned how the basics of nodes and graph workings and how to create them. In the code we used we still haven't used llm to decide which node to choose. For now we used random number generation for outcome. The general idea was similar to neural networks.