Video 1 Edited code
In this video we revisited creating basic graph from module 1 simple graph and then added some more details to it. We learned to implement dataclass of python for concise syntax for creating classes that are primarily used on storing data. We then also added pydantic state so that we can catch errors when passing keys to our state on runtime.