Video 2 Edited code
In this video we learned to update value of nodes on same level by appending instead of overwriting as that leads to problems. This was achieved by using reducers. Then we used custom reducers to handle edge cases. We then messagestate. We then did rewriting and removal of messages using id. This is very useful as we can now have full control of what stays in memory.