Video 5 Edited code
In this video we learned to set up summary where it will save summary of the messagestate and we can control on how many messages should the summary start. We can also do this while using thread so the messages doesn't have to be appended at same time for it to be summarized. The summary lives in its own state.