Video 6 Edited code
In this video we used Sqlite to set up external memory so that the lifetime of the saved messages and other states were not just the lifetime of the notebook kernel. Now even if we close the kernel we were able to verify that the information was stillsaved through Sqlite. W e then also viewed the summarization working in langgraph studio.