Video 2 Edited code
In this video titled breakpoints we get introduced to human in the loop by using breakpoints. There are 3 main uses for human in the loop which are for approval from user, debugging the code and editing the graph state. Breakpoints can be implemented so that approval is needed for the execution to continue from one node to next. This breakpoint can be implemented at tool call etc.