Video 5 Edited code
In this video we visited the last use of human in the loop debugging. We first went over some actions such as browsing history of state, replaying the states of the graph and forking where run same step but with different input. In forking instead of appending when updating state we replaced the input by using id. Then went over the same actions by langgraph api. So when we use these actions together we can go back to checkpoints and run from any checkpoint we want achieving time travel. This can also be done in langraph studio in an easy manner.