Video 2 Edited code
In this video I learned about how we can make graphs into subgraphs of a another graph thereby making it parent graph. In the example code we created the graphs question summarization and failure analysis and used them as subgraphs. The subgraphs run in parallel here. We use subgraphs when we want different states in our overall graph.