Video 3 Edited code
In this video we looked at map reduce which is a technique to implement parallelization to increase efficiency. This can be seen as two parts that is first divide a task into smaller task and second part is coming to the result from all the completed tasks. First part is called map and second part is called reduce. In the how we used this was we parallelly generated jokes for different topics. My edited code used it a little differently. As a side note we saw how we can create conditional edge also.