Correct Video 1 code
In this video I learned about how run nodes parallelly in the same step ie fanning out from one node then fanning in in a node in the next step. If we just build the graph and try execute the code having parallelly running node in a step it will return error but we can work around this by adding reducer which appends the updates to list. Then we looked at how we can set the order of these node execution in same step by adding reducer. Then we looked on an example of how we can implement this practically by doing web search and search Wikipedia parallelly.