Video 4 Edited code
In this video we looked at how we can create a research assistant. First of all a research assistant is preferred over general llm in cases where we want source to be specific or when the output should be in particular format. In this code we created a simulation where there is one expert and multiple interviewers. The sources related to the question were made available just to the expert where the sources are Wikipedia search and Web search run parallelly and appended to list(otherwise it would result in error as we learned in vid 1) from the question by interviewer. We are using human in the loop to select analysts. We also use map reduce to run these interviews is parallel. Here it becomes clear all the usefulness of the tools we learned earlier.